+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Add support for removing IAM permissions from the capa controller role when EKS support is not required.
+- Add support for removing IAM permissions from the capa controller role when resolver rules support is not required.
 
 ## [5.1.0] - 2025-02-20