+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [5.2.0] - 2025-03-25
+
 ### Added
 
 - Allow to override the principal ARN to assume the capa controller role.
@@ -191,7 +193,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - First release.
 
-[Unreleased]: https://github.com/giantswarm/giantswarm-aws-account-prerequisites/compare/v5.1.0...HEAD
+[Unreleased]: https://github.com/giantswarm/giantswarm-aws-account-prerequisites/compare/v5.2.0...HEAD
+[5.2.0]: https://github.com/giantswarm/giantswarm-aws-account-prerequisites/compare/v5.1.0...v5.2.0
 [5.1.0]: https://github.com/giantswarm/giantswarm-aws-account-prerequisites/compare/v5.0.0...v5.1.0
 [5.0.0]: https://github.com/giantswarm/giantswarm-aws-account-prerequisites/compare/v4.3.1...v5.0.0
 [4.3.1]: https://github.com/giantswarm/giantswarm-aws-account-prerequisites/compare/v4.3.0...v4.3.1